requirements.txt
discord.py
supabase
groq
python-dotenv
requests
